+++ packages/create-sst/bin/presets/graphql/dynamo/templates/services/core/dynamo.ts
@@ -2,10 +2,11 @@ export * as Dynamo from "./dynamo";
 
 import { EntityConfiguration } from "electrodb";
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
+import { Config } from "@serverless-stack/node/config";
 
 export const Client = new DynamoDBClient({});
 
 export const Configuration: EntityConfiguration = {
-  table: process.env.TABLE_NAME,
+  table: Config.TABLE_NAME,
   client: Client,
 };