+++ packages/resources/test/Function.test.ts
@@ -30,6 +30,7 @@ import {
   Function,
   FunctionProps
 } from "../src";
+import G from "glob";
 
 const lambdaDefaultPolicy = {
   Action: ["xray:PutTraceSegments", "xray:PutTelemetryRecords"],
@@ -431,17 +432,19 @@ test("xray-disabled", async () => {
 test("config", async () => {
   const stack = new Stack(new App(), "stack");
   const s = new Config.Secret(stack, "MY_SECRET");
+  const s2 = new Config.Secret(stack, "MY_SECRET2");
   const p = new Config.Parameter(stack, "MY_PARAM", {
     value: "value"
   });
   new Function(stack, "Function", {
     handler: "test/lambda.handler",
-    config: [s, p]
+    config: [s, s2, p]
   });
   hasResource(stack, "AWS::Lambda::Function", {
     Environment: {
       Variables: {
         SST_SECRET_MY_SECRET: "1",
+        SST_SECRET_MY_SECRET2: "1",
         SST_PARAM_MY_PARAM: "value",
         AWS_NODEJS_CONNECTION_REUSE_ENABLED: "1"
       }
@@ -457,6 +460,8 @@ test("config", async () => {
           Resource: [
             "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/MY_SECRET",
             "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/MY_SECRET",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/MY_SECRET2",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/MY_SECRET2",
           ],
         },
       ],
@@ -869,7 +874,7 @@ test("constructor: skipBuild", async () => {
 });
 
 /////////////////////////////
-// Test attachPermissions - generic
+// Test functions
 /////////////////////////////
 
 test("attachPermissions: string: all", async () => {
@@ -1422,6 +1427,54 @@ test("attachPermissions: policy statement", async () => {
   });
 });
 
+test("addConfig", async () => {
+  const stack = new Stack(new App(), "stack");
+  const s = new Config.Secret(stack, "MY_SECRET");
+  const s2 = new Config.Secret(stack, "MY_SECRET2");
+  const p = new Config.Parameter(stack, "MY_PARAM", {
+    value: "value"
+  });
+  const f = new Function(stack, "Function", {
+    handler: "test/lambda.handler",
+    config: [s, p],
+  });
+  f.addConfig([s2]);
+  hasResource(stack, "AWS::Lambda::Function", {
+    Environment: {
+      Variables: {
+        SST_SECRET_MY_SECRET: "1",
+        SST_SECRET_MY_SECRET2: "1",
+        SST_PARAM_MY_PARAM: "value",
+        AWS_NODEJS_CONNECTION_REUSE_ENABLED: "1"
+      }
+    }
+  });
+  hasResource(stack, "AWS::IAM::Policy", {
+    PolicyDocument: {
+      Statement: [
+        lambdaDefaultPolicy,
+        {
+          Action: "ssm:GetParameters",
+          Effect: "Allow",
+          Resource: [
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/MY_SECRET",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/MY_SECRET",
+          ],
+        },
+        {
+          Action: "ssm:GetParameters",
+          Effect: "Allow",
+          Resource: [
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/MY_SECRET2",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/MY_SECRET2",
+          ],
+        },
+      ],
+      Version: "2012-10-17"
+    }
+  });
+});
+
 /////////////////////////////
 // Test mergeProps
 /////////////////////////////
@@ -1777,6 +1830,54 @@ test("App.defaultFunctionProps(): permissions", async () => {
         lambdaDefaultPolicy,
         { Action: "s3:*", Effect: "Allow", Resource: "*" },
         { Action: "dynamodb:*", Effect: "Allow", Resource: "*" }
+      ],
+      Version: "2012-10-17"
+    }
+  });
+});
+
+test("App.defaultFunctionProps(): config", async () => {
+  const app = new App();
+  // Create a stack with 1 secret
+  const stackA = new Stack(app, "stackA");
+  const SECRET_A = new Config.Secret(stackA, "SECRET_A");
+  app.setDefaultFunctionProps({
+    config: [SECRET_A]
+  });
+
+  // Create another stack with 1 secret
+  const stackB = new Stack(app, "stackB");
+  const SECRET_B = new Config.Secret(stackB, "SECRET_B");
+  app.addDefaultFunctionConfig([SECRET_B]);
+
+  // Test function's config is merged with default config
+  const stack = new Stack(app, "stack");
+  const SECRET_C = new Config.Secret(stack, "SECRET_C");
+  stack.addDefaultFunctionConfig([SECRET_C]);
+  const SECRET_D = new Config.Secret(stack, "SECRET_D");
+  new Function(stack, "Function", {
+    handler: "test/lambda.handler",
+    config: [SECRET_D],
+  });
+
+  hasResource(stack, "AWS::IAM::Policy", {
+    PolicyDocument: {
+      Statement: [
+        lambdaDefaultPolicy,
+        {
+          Action: "ssm:GetParameters",
+          Effect: "Allow",
+          Resource: [
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/SECRET_A",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/SECRET_A",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/SECRET_B",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/SECRET_B",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/SECRET_C",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/SECRET_C",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/dev/secrets/SECRET_D",
+            "arn:aws:ssm:us-east-1:my-account:parameter/sst/my-app/.fallback/secrets/SECRET_D",
+          ],
+        },
       ],
       Version: "2012-10-17"
     }